Done exercise 4.20

+++ controllers/bloglist.js
@@ -7,7 +7,6 @@ import User from "../models/user.js";
 const blogRouter = Router()
 
 
-
 blogRouter.get("/", async (request, response) => {
 	const blogs = await Blog.find({})
 		.populate('user', {
@@ -21,7 +20,7 @@ blogRouter.post("/", async (request, response) => {
 	const body = request.body
 
 	const decodedToken = jwt.verify(request.token, process.env.SECRET)
-	if (!token || !decodedToken.id) {
+	if (!request.token || !decodedToken.id) {
 		return response.status(401).json({error: 'token missing or invalid'})
 	}
 
@@ -43,8 +42,19 @@ blogRouter.post("/", async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response, next) => {
-	await Blog.findByIdAndRemove(request.params.id)
-	response.status(204).end()
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!request.token || !decodedToken.id) {
+		return response.status(401).json({error: 'token missing or invalid'})
+	}
+
+	const blog = await Blog.findById(request.params.id)
+
+	if (blog.user.toString() === decodedToken.id.toString()) {
+		await Blog.findByIdAndRemove(request.params.id)
+		response.status(204).end()
+	}
+	response.status(401).end()
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
